# stellar-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar API
  description: |
    # Stellar API Documentation
    
    Modern REST API with JWT authentication, role-based access control, and comprehensive security features.
    
    ## Features
    - JWT-based authentication with refresh tokens
    - Role-based access control (RBAC)
    - Rate limiting and security headers
    - Input validation and sanitization
    - Comprehensive error handling
    - API versioning
    - Two-factor authentication support
    - Account lockout protection
    
    ## Authentication
    
    This API uses JWT (JSON Web Tokens) for authentication. To access protected endpoints:
    
    1. Register a new account or login with existing credentials
    2. Obtain an access token from the `/auth/login` endpoint
    3. Include the token in the `Authorization` header: `Bearer <token>`
    4. Use refresh tokens to obtain new access tokens when they expire
    
    ## Rate Limiting
    
    API requests are rate-limited to prevent abuse:
    - Default: 100 requests per 15 minutes per IP
    - Authentication endpoints: 5 requests per 15 minutes
    - Password reset: 3 requests per hour
    
    ## Error Responses
    
    All errors follow a consistent format:
    ```json
    {
      "error": "Error type",
      "message": "Human-readable error message",
      "statusCode": 400,
      "details": {} // Optional additional information
    }
    ```
    
  version: 1.0.0
  contact:
    name: Stellar Team
...
  license:
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          format: objectId
          example: "507f1f77bcf86cd799439011"
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: "^[a-zA-Z0-9_-]+$"
          example: "johndoe"
        email:
          type: string
          format: email
          example: "john@example.com"
        profile:
          $ref: '#/components/schemas/UserProfile'
        emailVerified:
          type: boolean
          example: true
        status:
          type: string
          enum: [active, inactive, suspended, pending]
          example: "active"
        role:
          type: string
          enum: [user, admin, moderator]
          example: "user"
        lastLogin:
          type: string
          format: date-time
        lastActive:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 50
          example: "John"
        lastName:
          type: string
          maxLength: 50
          example: "Doe"
        fullName:
          type: string
          readOnly: true
          example: "John Doe"
        avatar:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          maxLength: 500
          example: "Software developer passionate about building great APIs"
        dateOfBirth:
          type: string
          format: date
        phoneNumber:
          type: string
          pattern: "^\\+?[\\d\\s-()]+$"
          example: "+1-555-123-4567"
...
    UpdateProfileRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        profile:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
            bio:
              type: string
              maxLength: 500
            dateOfBirth:
              type: string
              format: date
            phoneNumber:
              type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "ValidationError"
        message:
          type: string
          example: "Invalid request parameters"
        statusCode:
          type: integer
          example: 400
        details:
          type: object
          additionalProperties: true
...
paths:
...
  /users/profile:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: Retrieve the authenticated user's profile information
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      summary: Update user profile
      description: Update the authenticated user's profile information
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...